[REFACTOR] - Cambio en la estructura del frontend, ejs, componentes
