+++ harvester-app/src/framework/utils/scripts/constantes.js
@@ -1,0 +1,17 @@
+// RUTA A SERVIDOR - ENTORNO MAIN
+// const URL_BASE='http://54.160.143.242'
+
+// RUTA A SERVIDOR - ENTORNO STAGING
+const URL_BASE='http://54.160.143.242/staging'
+
+// RUTA A SERVIDOR - ENTORNO LOCAL
+// const URL_BASE = 'http://localhost:3000';
+
+const LONGITUD_MAXIMA_NOMBRE_FORMULA=30
+const LONGITUD_MAXIMA_FORMULA=512
+
+module.exports = {
+	URL_BASE,
+	LONGITUD_MAXIMA_NOMBRE_FORMULA,
+	LONGITUD_MAXIMA_FORMULA
+}
