+++ harvester-app/src/framework/utils/scripts/middleware/auth.js
