+++ harvester-app/src/framework/utils/scripts/paginas/inicio/moduloTractores.js
