+++ harvester-app/src/framework/utils/scripts/pantallaCarga.js
