+++ harvester-app/src/framework/utils/scripts/sidebar.js
