+++ harvester-app/src/framework/vistas/includes/componentes/agregarGrafica.html
