+++ harvester-app/src/framework/vistas/includes/componentes/atomos/boton/boton.css
@@ -1,0 +1,7 @@
+btn-primario{
+    color : rojo;
+}
+
+btn-secundario{
+    
+}
