+++ harvester-app/src/framework/vistas/includes/componentes/atomos/boton/boton.ejs
@@ -1,0 +1,5 @@
+
+<!-- <%- include('../componentes/atomos/boton/boton.ejs', { label: 'Enviar', type: 'primario', id: 'boton1' }) %> -->
+
+<link href='boton.css' rel='stylesheet' />
+<button id='<%- id %>' class='btn btn-<%= type %>'><%= label %></button>
