+++ harvester-app/src/framework/vistas/paginas/contenedorPrincipal.ejs
@@ -1,0 +1,16 @@
+listaFormulas
+
+navbar
+
+<main>
+
+    sidebar
+
+    <principal>
+        listaFormulas
+    </principal>
+
+</main>
+
+
+footer
