+++ harvester-app/src/framework/vistas/paginas/pantallaCarga.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Cargando...</title>
+    <link rel="stylesheet" href="<%= basePath %>/framework/utils/css/pantallaCarga.css">
+</head>
+    <body>
+        <div class="loader"></div>
+        <script src="<%= basePath %>/framework/utils/js/pantallaCarga.js"></script>
+    </body>
+</html>
