+++ harvester-app/src/framework/vistas/paginas/usuarios/consultarPerfil.ejs
@@ -1,0 +1,1 @@
+<%- include('../components/button', { label: 'Enviar', type: 'primario', id: 'boton1' }) %>
