+++ harvester-app/src/framework/vistas/pantallaCarga.ejs
@@ -2,5 +2,5 @@
 
 <link href="/utils/css/prueba.css" rel="stylesheet"/>
 <h1 class = 'main-title'>Pruebaaa</h1>
-
+<%- nombre %> 
 <%- include('includes/footer.ejs') %>
